+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
+- Ensure that we pass floats through to the Float.\* functions, not integers.
 - Modify maths functions to accept both strings and numbers — strings will be coerced into numbers before applying the correct operation.
 - Add `roundTo` and `round_to` function as alias to `round` to mimic JS behaviour.